Former-commit-id: 962112a577015e38ec7b7edf85ee551a164dd878

+++ framework/sofa/core/objectmodel/Link.h
@@ -351,9 +351,9 @@ public:
     {
     }
 
-    unsigned int size(const core::ExecParams* params = 0) const
+    size_t size(const core::ExecParams* params = 0) const
     {
-        return (unsigned int)m_value[core::ExecParams::currentAspect(params)].size();
+        return (size_t)m_value[core::ExecParams::currentAspect(params)].size();
     }
 
     bool empty(const core::ExecParams* params = 0) const
@@ -458,7 +458,7 @@ public:
         return OwnerType::GetClass();
     }
 
-    unsigned int getSize() const
+    size_t getSize() const
     {
         return size();
     }
